Replace test_clone option with test_prepare
db:test:clone is deprecated because Rails is now supposed to handle that
automatically.

However, it doesn't handle redos, because it just compares schema and
migration timestamps. db:test:prepare, however, does handle this case,
is (no longer) deprecated, and is explicitly present in Rails for this
use-case: rails/rails#17739.